refactor(analysis-types): extract effort delta DTO
